Address joshs comments
